06_app/requirements.txt
streamlit==1.32.0
databricks-sql-connector==3.0.0
plotly==5.19.0
pandas==2.2.0
python-dotenv==0.21.0
fpdf2>=2.7.9